feat: Add onKeyDown event listener to text area in Upload page
This commit adds an onKeyDown event listener to the text area in the Upload page. When the Enter key is pressed, the add() function is called with the value of the text area. This allows users to submit the form by pressing Enter instead of clicking a button.

Note: This message includes additional information for clarity. Please remove it before using the commit message.