fix: Update author URL rendering in EntryPage component
This commit fixes an issue in the EntryPage component where the author URL was not being rendered correctly. The code change ensures that the author URL is displayed properly by parsing the metadata object and extracting the author value. This improves the user experience and provides accurate information to the users.