feat: Add cachedFData state to EntryPage component
This commit adds a new state variable, cachedFData, to the EntryPage component. The cachedFData state is used to store the previous value of fData before making changes to it. This allows for easy rollback to the last saved state in case of failed changes. The commit also includes minor code cleanup and commented out code for future reference.